requirements.txt
requests==2.32.3
beautifulsoup4==4.13.4
sentence-transformers==3.4.1
chromadb==1.5.5
groq
fastapi==0.115.12
uvicorn==0.34.0
python-dotenv==1.1.0
